Some changes.
Made a kind of HUD, which keeps track of the players' score and lives.
Replaced the mid-x and mid-y methods with fields.
Created an end-of-level method in game%, though it currently only exits
the game when there are no players or asteroids left.
Replaced all unnecessary init-fields with normal fields.
Added an obj-hash-collided-with method, because reasons... In the ship%
class obj-hash-collided-with will make the player invulnerable for a
short amount of time after collision. I did this because otherwise
(since ships have lives now and don't die on collision) four asteroids
would be created on collison with ships.